# Metaformer baselines for vision