PR adapting metaformers from https://github.com/sail-sg/metaformer to the main timm codebase.
Implements identityformer, randformer, poolformerv2, convformer, and caformer from https://arxiv.org/abs/2210.13452 and integrates poolformer (v1) from https://arxiv.org/abs/2111.11418.

Known issues:
- Randformer fails tests (not all parameters require gradients)
- Weights saved from old poolformer impl don't work
- Current checkpoint mapping is messy, should get changed to use `group_matcher`
- Excessive permute statements. Having everything channels-last was a workaround for torchscript. Should now be unnecesary following some other changes.
- Randformer only works on fixed input size (other models allow for arbitrary sizes)

I'd like to hear any additional feedback before proceeding to address these issues. My validation results line up with the paper's.