@HuggingFaceDocBuilderDev 2023-02-01
The docs for this PR live [here](https://moon-ci-docs.huggingface.co/docs/timm/pr_1647). All of your documentation changes will be reflected on that endpoint.