@fffffgggg54 2023-02-05
> I've converted models with unused params before, have you identified what the params are and why the authors might have not realized they weren't used? they should get removed for a cleaner model
> 
RandFormer has a matrix that maps input tokens to output tokens at random. It doesn't require grad, since it isn't meant to be updated during training but should still be included in saved checkpoints. The module is at https://github.com/fffffgggg54/pytorch-image-models/blob/metaformer_baselines_for_vision/timm/models/metaformers.py#L154.

For the checkpoint mapping, the current scheme works for pretrained weights that are loaded during the `timm.create_model()` call. My concern is that, in order to merge poolformer (v1) into this new impl, old checkpoints from the current poolformer impl that are saved by the user need to be mapped to the new model. Is there a way to support this or does the old impl need to stay for legacy support?
Ex:
```python
pf_original = timm.create_model('poolformer_s12')
torch.save(pf_original.state_dict(), 'original.pth')

# Doesn't work
pf_new = timm.create_model('poolformerv1_s12')
pf_new.load_state_dict(torch.load('original.pth'))
```