@rwightman 2023-02-22
@fffffgggg54 there's a lot going in this impl, some time to mentally parse

Looking at it, it looks 'slow' with so many permutes, the custom layer norm is **very** slow w/o using torch.compile (PyTorch 2.0) ass well.  I wonder if doing everything w/ convolutions would make more sense, linear tends to be faster as you scale up the width but these models aren't super large. Hmm.

For the random mixing, that should be a buffer not a parameter since it looks like it's supposed to be fixed.