@fffffgggg54 2023-03-06
Changed to channels first, randformer uses buffer and allows for arbitrary input sizes, still use einsum and run feature extraction test for randformer.