@fffffgggg54 2023-03-12
draft for now, convformer and caformer completely fail validation (~55% and ~0.02%), other 4 variants work.