@fffffgggg54 2023-03-15
The regression after the channels-first change had to do with the norm layers used. This has been fixed. The BCHW randformer implementation was also updated to match val results from paper, problem had to do with reshaping. I added an interpolationg op in the random mixing module. The model has the same val results at when run at train res, but getting actual performance out of it beyond slight variations and reductions, in eval size still wip. Change allows for varied input sizes should someone choose to do so with the model. Some other parts of the implementation may still need a bit of cleanup.

@yuweihao do you have any input? I know you mentioned wanting to have the new poolformer impl replace the old one, with the model renamed to `poolformer_*`. This would break backwards compatibility with weights saved from the old model definition, but since there is a new version of timm planned with backwards-compatability-breaking changes, I'm not sure if this is an issue.

<details><summary>validation results of channels-first are within error of ea054117ee65bd0ed1da16cbd7ffb7e15c3cf521 and the paper.</summary>

```
new:
model: identityformer_s12.sail_in1k, top-1: 74.58%, spent 721 seconds
model: randformer_s12.sail_in1k, top-1: 76.63799999999999%, spent 721 seconds
model: poolformerv1_s12.sail_in1k, top-1: 77.14999999999999%, spent 721 seconds
model: poolformerv2_s12.sail_in1k, top-1: 77.996%, spent 721 seconds
model: convformer_s18.sail_in1k, top-1: 82.98%, spent 721 seconds
model: caformer_s18.sail_in1k, top-1: 83.66799999999999%, spent 721 seconds

old:
model: identityformer_s12.sail_in1k, top-1: 74.58%, spent 699 seconds
model: randformer_s12.sail_in1k, top-1: 76.63%, spent 699 seconds
model: poolformerv1_s12.sail_in1k, top-1: 77.146%, spent 699 seconds
model: poolformerv2_s12.sail_in1k, top-1: 78.006%, spent 699 seconds
model: convformer_s18.sail_in1k, top-1: 82.98400000000001%, spent 699 seconds
model: caformer_s18.sail_in1k, top-1: 83.662%, spent 699 seconds
```
</details>

<details><summary>nn.Linear ->1x1 conv results in noticeable (17-20%) performance gain</summary>

```
identityformer_s12, tesla m40, bs=64, 1x1 conv: ~155.5 ips
identityformer_s12, tesla m40, bs=64, fc: - ~129.1 ips

identityformer_m48, tesla m40, bs=16, 1x1 conv: - ~27.6 ips
identityformer_m48, tesla m40, bs=16, fc: - ~23.45 ips
```
</details>

<details><summary>224 train -> new test resolution, randformer interpolation performs poorly, other models as expected</summary>

```
eval @ 128:
model: identityformer_s12.sail_in1k, top-1: 63.94800000000001%, spent 455 seconds
model: randformer_s12.sail_in1k, top-1: 64.862%, spent 455 seconds
model: poolformerv1_s12.sail_in1k, top-1: 65.684%, spent 455 seconds
model: poolformerv2_s12.sail_in1k, top-1: 67.22%, spent 455 seconds
model: convformer_s18.sail_in1k, top-1: 75.8%, spent 455 seconds
model: caformer_s18.sail_in1k, top-1: 75.68%, spent 455 seconds

eval @ 160:
model: identityformer_s12.sail_in1k, top-1: 69.652%, spent 528 seconds
model: randformer_s12.sail_in1k, top-1: 71.544%, spent 528 seconds
model: poolformerv1_s12.sail_in1k, top-1: 72.25200000000001%, spent 528 seconds
model: poolformerv2_s12.sail_in1k, top-1: 73.262%, spent 528 seconds
model: convformer_s18.sail_in1k, top-1: 79.852%, spent 528 seconds
model: caformer_s18.sail_in1k, top-1: 80.31%, spent 528 seconds

eval @ 288:
model: identityformer_s12.sail_in1k, top-1: 75.47399999999999%, spent 1022 seconds
model: randformer_s12.sail_in1k, top-1: 75.31400000000001%, spent 1022 seconds
model: poolformerv1_s12.sail_in1k, top-1: 78.082%, spent 1022 seconds
model: poolformerv2_s12.sail_in1k, top-1: 79.30199999999999%, spent 1022 seconds
model: convformer_s18.sail_in1k, top-1: 83.572%, spent 1022 seconds
model: caformer_s18.sail_in1k, top-1: 83.72%, spent 1022 seconds

eval @ 384:
model: identityformer_s12.sail_in1k, top-1: 74.76%, spent 1815 seconds
model: randformer_s12.sail_in1k, top-1: 1.8599999999999999%, spent 1815 seconds
model: poolformerv1_s12.sail_in1k, top-1: 77.63799999999999%, spent 1815 seconds
model: poolformerv2_s12.sail_in1k, top-1: 78.956%, spent 1815 seconds
model: convformer_s18.sail_in1k, top-1: 82.178%, spent 1815 seconds
model: caformer_s18.sail_in1k, top-1: 81.582%, spent 1815 seconds
```
</details>