@yuweihao 2023-03-15
Hi @fffffgggg54 , many thanks for your contribution.

Could I give some small suggestions? It is better that your contributed `metaformers.py` does not contain old poolformer. That is, it is better to remove poolformerv1 in `metaformers.py`. This way may make the code have better compatibility. 

Besides, it is better that `metaformers.py` is renamed to `metaformer_baselines.py`. The first reason is that metaformer is an abstract concept while metaformer baselines are specific models. The second reason is that it can remind users that old poolformer is in `poolformer.py` and poolformerv2 is in `metaformer_baselines.py`.

Thank you very much again for your contribution.